+++ app/controllers/editor/editor_controller.rb
@@ -24,22 +24,4 @@ class Editor::EditorController < ApplicationController
   def redirect_without_team
     empty_response_or_root_path(editor_project_ae_teams_path(@project)) unless @team
   end
-
-  def find_filter_or_redirect(id = :check_filter_id)
-    @check_filter = @check.check_filters.find_by(id: params[id])
-    redirect_without_filter
-  end
-
-  def redirect_without_filter
-    empty_response_or_root_path(editor_project_check_path(@project, @check)) unless @check_filter
-  end
-
-  def find_filter_value_or_redirect(id = :check_filter_value_id)
-    @check_filter_value = @check_filter.check_filter_values.find_by(id: params[id])
-    redirect_without_filter_value
-  end
-
-  def redirect_without_filter_value
-    empty_response_or_root_path(editor_project_check_check_filter_path(@project, @check, @check_filter)) unless @check_filter_value
-  end
 end