+++ config/routes.rb
@@ -211,10 +211,6 @@ Rails.application.routes.draw do
         member do
           post :request_run, path: "request-run"
         end
-
-        resources :check_filters, path: "filters" do
-          resources :check_filter_values, path: "values"
-        end
       end
 
       resources :grid_variables, path: "grid-variables"