+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **Translation Changes**
   - French Canadian language added
   - Project editors can now specify the languages available to the project
+- **Variable Changes**
+  - Formats for calculated variables are now applied while editing a sheet
 - **Gem Changes**
   - Update to ruby 2.6.1
   - Update to rails 6.0.0.beta1